move HomePageLocalViewModel to LocalMusicViewModel redesign test for QuietMediaPlayer QuietMusicPlayer MusicPlayerManager
