+++ app/src/androidTest/java/tech/soit/quiet/player/core/QuietMediaPlayerTest.kt
@@ -6,8 +6,7 @@ import androidx.test.InstrumentationRegistry
 import kotlinx.coroutines.experimental.delay
 import kotlinx.coroutines.experimental.runBlocking
 import org.junit.After
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertTrue
+import org.junit.Assert.*
 import org.junit.Before
 import org.junit.Test
 
@@ -42,7 +41,7 @@ class QuietMediaPlayerTest {
         assertTrue(quietMediaPlayer.getState() == IMediaPlayer.IDLE)
         quietMediaPlayer.prepare(URI, true)
         assertTrue(quietMediaPlayer.getState() == IMediaPlayer.PREPARING)
-        delay(100)
+        delay(1000)
         assertTrue(quietMediaPlayer.getState() == IMediaPlayer.PLAYING)
     }
 
@@ -51,15 +50,15 @@ class QuietMediaPlayerTest {
         quietMediaPlayer.prepare(URI, false)
         assertTrue(quietMediaPlayer.getState() == IMediaPlayer.PREPARING)
 
-        delay(100)
+        delay(1000)
         assertTrue(quietMediaPlayer.getState() == IMediaPlayer.PAUSING)
 
         quietMediaPlayer.isPlayWhenReady = true
-        delay(100)
+        delay(1000)
         assertTrue(quietMediaPlayer.getState() == IMediaPlayer.PLAYING)
 
         quietMediaPlayer.isPlayWhenReady = false
-        delay(100)
+        delay(1000)
         assertTrue(quietMediaPlayer.getState() == IMediaPlayer.PAUSING)
 
     }
@@ -67,14 +66,14 @@ class QuietMediaPlayerTest {
     @Test
     fun seekTo() = runBlocking {
         quietMediaPlayer.prepare(URI, true)
-        delay(100)
+        delay(1000)
         assertTrue(quietMediaPlayer.getState() == IMediaPlayer.PLAYING)
 
-        assertTrue(quietMediaPlayer.getPosition() < 1000)
+        assertTrue("current position(${quietMediaPlayer.getPosition()}) less than 2000", quietMediaPlayer.getPosition() < 2000)
 
-        quietMediaPlayer.seekTo(3000)
-
-        assertTrue(quietMediaPlayer.getPosition() > 2500)
+        quietMediaPlayer.seekTo(4000)
+        delay(100)
+        assertTrue("current position(${quietMediaPlayer.getPosition()} greater than 3500", quietMediaPlayer.getPosition() > 3500)
 
     }
 
@@ -83,22 +82,23 @@ class QuietMediaPlayerTest {
     fun getDuration() = runBlocking {
         assertEquals(quietMediaPlayer.getDuration(), 0L)
         quietMediaPlayer.prepare(URI, false)
-        delay(100)
+        delay(1000)
         assertTrue(quietMediaPlayer.getState() == IMediaPlayer.PAUSING)
 
-        assertTrue(quietMediaPlayer.getDuration() > 5000)
+        assertTrue("duration ${quietMediaPlayer.getDuration()} greater than 7000", quietMediaPlayer.getDuration() > 7000)
+        assertTrue("duration ${quietMediaPlayer.getDuration()} less than 9000", quietMediaPlayer.getDuration() < 9000)
     }
 
     @Test
     fun testPlayOther() = runBlocking {
 
         quietMediaPlayer.prepare(URI, true)
-
-        delay(20)
+        delay(1000)
+        assertEquals("player is playing", IMediaPlayer.PLAYING, quietMediaPlayer.getState())
 
         quietMediaPlayer.prepare(URI, true)
-        delay(100)
-        assertTrue(quietMediaPlayer.getState() == IMediaPlayer.PLAYING)
+        delay(1000)
+        assertEquals("play an other music is still playing !!", IMediaPlayer.PLAYING, quietMediaPlayer.getState())
 
     }
 
@@ -106,15 +106,17 @@ class QuietMediaPlayerTest {
     fun testRelease() = runBlocking {
 
         quietMediaPlayer.prepare(URI, true)
-        delay(20)
-        assertTrue(quietMediaPlayer.getState() == IMediaPlayer.PLAYING)
+        delay(1000)
+        assertEquals("player is playing..", IMediaPlayer.PLAYING, quietMediaPlayer.getState())
 
         quietMediaPlayer.release()
-        assertTrue(quietMediaPlayer.getState() == IMediaPlayer.IDLE)
+
+        assertEquals("player has been release", IMediaPlayer.IDLE, quietMediaPlayer.getState())
+        assertFalse("player has been release", quietMediaPlayer.isPlayWhenReady)
 
         quietMediaPlayer.prepare(URI, true)
-        delay(100)
-        assertTrue(quietMediaPlayer.getState() == IMediaPlayer.PLAYING)
+        delay(1000)
+        assertEquals("player replay after release", IMediaPlayer.PLAYING, quietMediaPlayer.getState())
 
     }
 }
